backend/requirements.txt
# python 3.8.8 (old right? coz experta doesn't support 3.9+)

experta==1.9.4
ipython==7.34.0
django==4.2.28
django_ninja>=1.0
django_cors_headers>=4.0
uvicorn==0.33.0
httpx==0.28.1
python_dotenv==1.0.1